bug: 18787 It is a div element style overflow issue. To fix it, we set the ListView's overflow to auto and list row's overflow to visible. In this way, the list row content can always be visible and the scroll bar of the list view can control the visible area of the list view contents.
At the same time, we change the auto size of the last column of the list into an arbitrary size instead of auto. When the size is autoed, the last column content will not display if the list view size is small because it is automatically reduced to 0. 

We also move the setScrollStyle to the parent class of all the listview - ZaListView.js.

http://bugzilla.zimbra.com/show_bug.cgi?id=18787