bug: 6082
First pass at notification on delegate stores.  The user still has only 1 SoapSession, but when access to a delegate store is requested we spawn a listener on that mailbox as well.  This sub-session is not registered in the cache, so it doesn't time out on its own and doesn't have an externally addressable session ID.  When the parent session is reaped, the delegate sessions are unregistered as listeners as well.

Notifications fed to delegate sessions are massaged (e.g. items moved to non-viewable folders are treated as deletes) and filtered and then stored in the main SoapSession's notification queue.  A side effect of this is that the keys of the PendingModification hashes must now be qualified IDs (represented as accountID, itemID Pairs).

Works only on delegate folders on the same host as the user's mailbox, and works for all ops other than searches.  There's an issue with conversation notifications being dropped, but notifications of changes to messages, contacts, appointments, etc. should be fine.

http://bugzilla.zimbra.com/show_bug.cgi?id=6082