Bug: 10118
Simplify the handling of auth-related exceptions.

We have always had special handling for AUTH_EXPIRED, so
that after the user re-auths, the request can be re-tried without
user intervention. Each caller of sendRequest() had to set up
a callback for that to happen, and a minority of callers did that.

Now it happens automatically for all async requests. The exception
handler saves the original request params, and hands them back
to sendRequest() after the user re-auths. Sync requests still need
to specify a callback since they have state to preserve.

The term "execFrame" no longer exists in our client code :)

Thanks to Dave C for the idea to do this!

http://bugzilla.zimbra.com/show_bug.cgi?id=10118