bug: 22873
When storing the blob bytes in redo op to file during redo, ignore the path in the redo op and save to a newly assigned incoming file path, if the redo is for a single-recipient delivery.  This blob is later renamed to the final path under target mailbox's blob directory.  If we saved the blob to the path in the redo op, we would end up moving the file from one mailbox to another when the redo is being done with a different target mailbox, as in restore to a new account address.

http://bugzilla.zimbra.com/show_bug.cgi?id=22873