Bug: 14156
A few changes that seem to improve speed in IM (hold a reference to the DOM element in each widget, rather than calling document.getElementById all the time; got rid of Dwt.toDocumentFragment in DwtToolBar since it's quite heavy and called for each widget added to a toolbar).

These changes should be an improvement for all other apps as well.

http://bugzilla.zimbra.com/show_bug.cgi?id=14156