Bug: 22234 Drag organizer over tag tree -> JS error
Rewrite how drag sources and targets are matched up.
The problem stemmed from the fact that we use the class
of the dragged object (via instanceof) to check to see if it's
valid for the dragged-over target. If that class has not been
loaded, we get a JS error.

Now, it happens in two steps. To add an association, you only
need to worry about the first step: when registering an item or
an organizer, you list the valid drop targets (note that this is the
reverse of the way we used to do it, which was to have the drop
targets declare what could be dragged onto them).

The second step happens when the package containing the item and
organizer classes is loaded (we assume it's the app's core class). The
droppable objects are actually registered with the appropriate drop
targets (all organizers) at this point, since the classes have been loaded.

* ZmTreeController's constructor no longer takes a dropTgt
* ZmTreeController._getDropTarget is gone
* All drag sources now have a complex data object with props for
   the data and the controller (that was previously true only for items)

http://bugzilla.zimbra.com/show_bug.cgi?id=22234