bug: 23601
Next alarm computation logic was ignoring task instances whose start time is earlier than the current value of next alarm time.  This check was a short circuit to reduce the number of instances to examine.  It works for appointments because they always have a valid instance start time, but tasks may not have DTSTART set, and this was represented as 0 value.  With the next alarm time value being a positive number, this resulted in ignoring all instances and clearing the next alarm from the task item.

Solution is to _not_ ignore the instance if it's a task instance that didn't have DTSTART set.

http://bugzilla.zimbra.com/show_bug.cgi?id=23601