bug: 23886 - Run spam test before user filters during message delivery
Updated RuleManager to not apply filter rules to spam by default.

Added the zimbraSpamApplyUserFilters attribute to account/cos, which allows
a sysadmin to revert a user to the old behavior.  This attribute is put in as
a safety measure for backward compatibility and should be pulled out in
a future release to avoid confusion with differences in filter behavior.

Added a unit test that verifies how spam is affected by user filters.

http://bugzilla.zimbra.com/show_bug.cgi?id=23886