bug: 23774
When there are system-level mutators or when an Outlook iCal fixup would actually change the message, don't use the workaround from bug 19028.  This means that malformed parts would cause send to fail, but this is better than the alternative (ZCO delegated send produces malformed messages).

When we change out the MIME parser, this all becomes moot.  I <heart> JavaMail.

http://bugzilla.zimbra.com/show_bug.cgi?id=23774