bug: 22594
Changes necessary to allow attendees (non-organizers) to set local reminders

When organizer sends an updated invite, the reminders from local invite is carried over to the new invite, thus preserving the attendee's reminders.  If attendee sets reminders on an instance of a recurring appointment, a local exception instance is created.  This model is compatible with Outlook behavior.  If the organizer then updates the recurring series, the exception is recreated by merging the data from the series update invite and the reminders from the previous exception.  This allows the attendee to see the latest published data, plus his/her own reminders.

An important change of behavior had to be made when updating the series of recurring appointment.  Previously, we allowed exceptions to remain.  For instance, suppose there is a recurring appointment every Monday at noon, except the 2nd week's instance is at 1pm.  If series is updated to start at 2pm instead, the 1pm instance would remain at 1pm.  Any canceled instances would remain canceled too.  But the new behavior blows away all exceptions.  After updating the series to 2pm start time, the 2nd week instance would also start at 2pm, and all previously canceled instances would be resurrected.

This is a somewhat surprising behavior, but it's necessary to go to this model to remove the ambiguity for updating existing exceptions.  When processing the series update from the organizer and attendee's appointment has an exception, it's unclear if that exception is also present in the organizer's appointment (in which case we should leave the exception) or if it's present only in the attendee's copy, perhaps created to define reminders (in which case we should recreate the exception with series data plus existing reminders).  Having the policy that series update always invalidates all existing exceptions gets us out of the dilemma.  Outlook also behaves this way.

http://bugzilla.zimbra.com/show_bug.cgi?id=22594