bug: 19222
Split MimeHeader into its own separate class.

Make a number of MimePart setter methods return "this" instead of returning void so that multiple calls can be chained.

Tweak the quoted-printable encoder/decoder to share some code and to allow encoders to specify additional characters that must be qp-encoded.

Slightly change the behavior of the ContentType parser when it hits unexpected CR or LF.

http://bugzilla.zimbra.com/show_bug.cgi?id=19222