bug: 24118
Before doing a cross-mailbox move/copy, check to see if the target folder is a mountpoint and dereference as necessary.  (This is more expensive than just blindly doing a delete/create, but it's more correct and I think it'll average out about equal in the long term due to optimizing out moves from a subfolder back to the mountpoint.)

Track dereferencing hopcount and max at 5 hops.

http://bugzilla.zimbra.com/show_bug.cgi?id=24118