Bug: 24327 - Eliminate expando properties in DOM elements
Replace DwtUiEvent.getDwtObjFromEvent() with
DwtControl.getTargetControl(), which works its way up from
the event target until it finds an element with an ID that maps
to a control. The former method is now deprecated.

A perf test showed about a 30% gain in speed. This is the first
step in eliminating the need for the "dwtObj" property in DOM
elements.

http://bugzilla.zimbra.com/show_bug.cgi?id=24327