Bug: 24237 - Eliminate expando properties in DOM elements
Changes to DwtUiEvent.setFromDhtmlEvent, so that it no longer
grubs through the DOM to set the event's dwtObj. Most of the
time, that's not needed, and some of the time it was redundant.
The function now takes an optional second arg: obj. If provided,
it is used to set dwtObj. If it is set to true, then the target control
will be fetched from the DHTML event and used to set dwtObj.

http://bugzilla.zimbra.com/show_bug.cgi?id=24237