Bug: 24327 - Eliminate expando properties in DOM elements
Rework how list views track items/elements, after removing a bunch
of expando properties from the elements. Now we use an element's
ID (which is tied to the item it represents) as a key into a hash to get
info that used to be in the element.

The main expando props that are gone are:

	_itemIndex
	_type

Calls to AjxCore.objectWithId() have been replaced. The following
methods are now available in DwtListView to manage item/element relations:

associateItemWithElement(item, el, type, id, data)
getItemFromElement(el)
findItem(el)
findItemDiv(el)
getTargetItem(ev)
getTargetItemDiv(ev)
_getItemData(el, field)
_setItemData(id, field, value)

http://bugzilla.zimbra.com/show_bug.cgi?id=24327