bug: 25051
For PreAuthServlet in the proxy-case(zimbraMailReferMode == never), treat the non-home server case the same as the home server case.  That is, always set auth cookie and redirect to the relative mail URL, regardless if the request is on the right server.

e.g. if the user types in the proxy URL:  http://mail.company.com/
and the proxy forwards them to a random host (only visible internally), like mail99.company.com, and it does a relative redirect to the mail URL, then they still end up at http://mail.company.com, which is the desired behavior.

http://bugzilla.zimbra.com/show_bug.cgi?id=25051