bug: 25095
Rework UUENCODE decoder to use InputStreams instead of Strings.  Move the abort callback up to the point where the first extractable attachment is parsed out.

Note that when messages with UUENCODEd attachments are converted, the resulting extracted attachments and trimmed main body part are stored in memory rather than on disk.  We can cosider writing this data to a temp "upload" file, but at present the auto-expiry of such files would cause the message cache to break.

http://bugzilla.zimbra.com/show_bug.cgi?id=25095