bug: 26113
For ModifyAccount, if the account's home server is changed as a result of this command and the new server is no longer this server, send a flush cache command to the new server so we don't get into the following:
     - account is on server A (this server)
     - on server B, do 
           zmprov ma {account} zimbraMailHost B
               (the ma is proxied to server A;  and on server B, the account still appears to be on A)
           then 
           zmprov ma {account} {any attr} {value}
           will get error:  ERROR: service.TOO_MANY_HOPS
           Until the account is expired from cache on server B.      

If the flushCache failed for any reason(e.g. the new server is not running), we log a warning, the ModifyAccount command is still successful. 

Without this fix, the above scenario can be worked around by manually  issuing a "zmprov flushCache account {acct id or name}" to the new home server of the account after modifyAccount.

http://bugzilla.zimbra.com/show_bug.cgi?id=26113