bug: 25381 Read the zimbraPrefLocale from GetInfoResponse, and reload all the message files if zimbraPrefLocale present and different from the default locale.
In order to make sure that no old locale message is used, we should prevent any widget to be created before the message reload. However, ZaController has a few dialogs which are created before the message reload, so we recreate those error dialog and message dialog after the message reload.

Please note, the way we handle the zimbraPrefLocale is different from ZCS Client which handles it  on the server side in JSP files. Admin Console handles it on the client side. It is due to the fact that all the javascripts and message files are loaded before user login, and we have no way to know what is user's locale preference until user logins. We may also handle the zimbraPrefLocale on the server side, however, it requires reload of the whole application which seems too expensive.

http://bugzilla.zimbra.com/show_bug.cgi?id=25381