bug: 2161 Y!!!!!!! Watch out..... complete new client side converter code eliminates all but a single class from the old convertd client/server and that one will be removed in main after I merge to Franklin (where we will use the old code by default)
I had to add a second convert() call to MimeHandler that doesn't require passing files in the filesystem. I would have had it return an InputStream to enable streaming, but text extraction requires a full string so the extra work/optimization is not worth it for now.

Also fixed an infinite loop case in the error msg redirection case and cleaned up zip archive and error html output

Marc - if you are reading this - Boris OWES me - he is a free man now...! :)

http://bugzilla.zimbra.com/show_bug.cgi?id=2161