bug: 27173, 27246
bug 27173 - BlobInputStream OOME during post-startup crash recovery

Restructured collection of substreams so that they are stored in
a single set instead of hierarchically.  When a stream is closed,
it is now removed from the set so that it can be garbage collected.

bug 27246 - Eliminate finalizers where possible

Removed BlobInputStream.finalize().  Should be safe.  The only
possible fallout is that a substream that is not closed will hang
around until the MimeMessage is garbage collected.  

http://bugzilla.zimbra.com/show_bug.cgi?id=27173
http://bugzilla.zimbra.com/show_bug.cgi?id=27246