bug: 21442
Add <IMSetIdleRequest> SOAP api:

<IMSetIdleRequest isIdle="1|0" idleTime="seconds_idle"/>
<IMSetIdleResponse/>

-- idleTime is number of seconds since last user action.  Client should
make a best-effort to calculate this, but is free to return "0" if
unsupported.

-- Throws ServiceException.SESSION_REQUIRED if there is no session
   referenced (a session MUST be active for this API to be used)

When calculating presence, the sever will look at the "idle" values of
all active web client sessions.  If all sessions are "idle" then the
server will update the user's status to "away".

The web client must be careful to unset the "idle" status when the
user comes back, otherwise the user's presence will be stuck in the
"Away" state and the user won't understand why.

http://bugzilla.zimbra.com/show_bug.cgi?id=21442