Integrate change 95207 from FRANKLIN to main
-------------
Bug: 26517 - Clicking on Next or previous button is displaying the second next message to the current message or the second previous message to the current message
This was happening because the mouseup event was hitting
two handlers. The root cause is that the original mouse event was
getting overwritten as part of pagination (when we simulate a
dbl-click), and stopPropagation's true value was cleared. That
has been fixed - we no longer reuse DwtShell.mouseEvent when
simulating a mouse event.
Toolbars handle events as a whole, so it's not necessary for their
components to add handlers. The other part of this bug fix is to
handle the case where a toolbar contains another toolbar.
- also, remove unused args to DwtToolBar ctor

http://bugzilla.zimbra.com/show_bug.cgi?id=26517