Integrate change 96313 from FRANKLIN to main
-------------
Bug: 16337 - Hybrid: would like choice whether highlighting a msg marks it 'Read'
Give user control over when a msg is marked read by the client. Choices are:
* Mark it read immediately
* Mark it read after N seconds
* Do not mark it read
That preference applies to messages that appear in the reading pane in a
double-pane view. If the message is opened in MV (double-clicked) or into
a detached window, it is always marked read.
If the user wants the msg marked read after a certain amount of time,
the reading pane must contain the msg for that time. If another msg is
displayed or if the user switches to another view during that time, the
msg is not marked read. The timer is only started the first time the
user views the msg; it is not restarted if the user leaves the view and
comes back.
- create pref MARK_MSG_READ
- paramify ZmMailMsg.load() and ZmSearch.getConv()

http://bugzilla.zimbra.com/show_bug.cgi?id=16337