bug: 27843, 6082
When expanding mountpoints pointing at folders on the same box, use an ItemIdFormatter appropriate to the target mailbox.  Also, serialize and cache the mounted folder structure immediately, then merge the serialized version of the mountpoint targets into the serialized local mailbox.  The latter unifies the logic for handling local-target and remote-target mountpoints.

http://bugzilla.zimbra.com/show_bug.cgi?id=27843
http://bugzilla.zimbra.com/show_bug.cgi?id=6082