bug: 18385, 6937
Zimbra authenticator is now known by mechanism "X-ZIMBRA".  Its initial (and only) response has been changed to be identical to that of AUTH PLAIN, except that X-ZIMBRA takes a valid Zimbra auth token as the password (e.g. "authzid^@authcid^@authtoken").  The authtoken must match authcid, and then the normal authorization check is made for authzid with the caveat that admin access is granted only if the supplied authtoken is an admin token.

When handling LOGIN and AUTH PLAIN, put the requested username in the auth context map.

More concise error logging for auth errors.

http://bugzilla.zimbra.com/show_bug.cgi?id=18385
http://bugzilla.zimbra.com/show_bug.cgi?id=6937