bug: 6937
Support SELECT/EXAMINE of folders on other servers.  This is done by using the new IMAP client to open a new connection and authenticate via 
AUTHENTICATE X-ZIMBRA, and then thereafter funneling the SELECT and all subsequent SELECT-dependent requests over to the remote sevrer.  Some requests (LIST, APPEND, SETACL, SUBSCRIBE, etc.) are still satisfied without proxying, but in those cases pending notifications are still fetched from the remote server via NOOP and returned inline with the response.  The proxy is valid for the duration of a SELECT/EXAMINE and is dropped when the folder is deselected.  IDLE is not supported.

More concise logging of auth failures.  Made a few classes and constructors package-private.  Cleaned up dropConnection() to use unsetSelectedFolder() (which now drops the proxy connection as well).

http://bugzilla.zimbra.com/show_bug.cgi?id=6937