bug: 27369
return dummy domain name and id when we fallbaclk to global config, becasue we don't really have a domain to get name/id for.   Unless we chaneg the SOAP API, client will be choked when it tries to get the required SOAP attrs name and id from the <domain> element. 

http://bugzilla.zimbra.com/show_bug.cgi?id=27369