bug: 28309
Exclude "fileAs" from the general contact search body, that way searches like "1" or "2" won't match the fileAs string.  I suspect this bug will come back -- we'll probably need some kind of a system for marking which contact fields are searchable as general data, and which ones are for not, but for now I'm only excluding fileAs.

Fixed a problem with index updating -- we weren't correctly deleting the old indexed data in some cases when re-indexing mutable mail_items.

http://bugzilla.zimbra.com/show_bug.cgi?id=28309