bug: 28830
p4 integrate -b FRANKLIN_to_main @104339,104339
make zdesktop server take a large attachment.  if remote server rejects the message due to size limit, we now catch that and put a notification message in user inbox.  in addition we move the outbound message back to drafts folder so that user can adjust it.

http://bugzilla.zimbra.com/show_bug.cgi?id=28830