bug: 29473, 29505
When updating tag or folder counts, make sure not to force a MOD_METADATA change unless requested.

Add a debug mode to automatically consistency-check modified items in the Mailbox's item caches at the end of a transaction.  If the localconfig key "debug_check_mailbox_cache_consistency" is set, we go to the database and do a field-by-field comparison to verify that the item's data in the cache exactly matches its row in the database.

http://bugzilla.zimbra.com/show_bug.cgi?id=29473
http://bugzilla.zimbra.com/show_bug.cgi?id=29505