bug: 29563
Support for derby as a ZCS database.  Just set the "ZIMBRA_USE_DERBY" environment variable and reset-the-world.  (Note that there are a few open issues with the IMAP server and derby, as that code was never needed for Desktop.)

Break IM tables off from db.sql into their own file so we can avoid loading them in Desktop.  Likewise, move the default volume initialization for reset-the-world into its own (shared) file.  Only copy the JDBC jarfile appropriate to the selected DB backend.

http://bugzilla.zimbra.com/show_bug.cgi?id=29563