Integrate change 106656 from FRANKLIN to main
-------------
Bug: 28850 - Cannot shift focus to message body
Make sure that a control really does have focus before
asking it to handle a key event.
The problem here is that the keyboard mgr's blur handler
does not null out the focus obj (if it did, then when the
browser loses and regains focus, the control that had
focus would no longer have or show it). So we use other
ways to see if the control actually has focus.
FF:
- can use Tab to give focus to msg view
- click in msg header, up/down arrow scrolls msg
- click in msg body, up/down arrow scrolls msg
IE:
- cannot use Tab to give focus to msg view
- click in msg header, up/down arrow does nothing
- click in msg body, up/down arrow scrolls msg

http://bugzilla.zimbra.com/show_bug.cgi?id=28850