bug: 25376 - Change "zmprov ral" to remove all account loggers
Updated SOAP API for RemoveAccountLoggerRequest so that the
user and category are optional.  Details of the new behavior are
in soap-admin.txt.

Updated all account logger soap
API's to specify the account as <account by="{name | id}">, like
our other API's.  The old <id> element is still supported, but logs
a deprecation warning when used.

Updated the ProvUtil removeAccountLogger command so that the
account and category are optional.

Added SOAP unit tests that validate the new functionality.

http://bugzilla.zimbra.com/show_bug.cgi?id=25376