bug:  28178
deleted zimbraReverseProxyAdminAccount/zimbraReverseProxyAdminAccountPassword from zimbra-attrs.xml, shift oids after them up.   They've never been released in any customer release.

from yunohu:   ;)
"As long as you are only changing the OIDs of an existing attribute,
OpenLDAP won't care.  The only time there would be a problem is if the
attribute existed in the DB and we got rid of it."

http://bugzilla.zimbra.com/show_bug.cgi?id=28178