bug: 30808
p4 integrate -b FRANKLIN_to_main @111275,111275
now that we consolidated two port into one, server code can mistake a regular request for admin request because the two ports are the same.  so the workaround is to set zimbraAdminPort server attribute to a bogus value since within desktop we shouldn't need a real value for this anyway.

http://bugzilla.zimbra.com/show_bug.cgi?id=30808