bug: 30640
SearchRequest for mail messages now supports headers. Anything in ZmMailMsg.requestHeaders are added tot he SearchRequest while fetching results.

This solves the bug of not getting GoodMail headers if the mail msg is the first on the list since we fetch the first mail as a part of SearchRequest of the mail folder.

http://bugzilla.zimbra.com/show_bug.cgi?id=30640