bug: 30002
cache non-existing domains so we don't repeatedly search LDAP for non-exsting domains.   Entries in the NonExistingCache has the same TTS/max as this DomainCache.    When a previously cached non-existing domain is created/found,  its entry in the non-existing cache is cleared and the entry is then cached normally.

zmprov fc domain {domain name/id} works on non-existing cache too.

http://bugzilla.zimbra.com/show_bug.cgi?id=30002