bug: 16472
p4 integrate -b FRANKLIN_to_main @113981,113981
Add a system folder called "Archive".  Everything moved under that folder are local only and not sync'ed with zimbra mailbox.  When an item or a tree of items are moved into anywhere under "Archive", we delete them from server.  When items are moved from archive to rest of desktop mailbox, we create those as new items on server.

http://bugzilla.zimbra.com/show_bug.cgi?id=16472