bug: 30938 Hard-coded Re: and Fwd: for reply/forward email compose.
However, I left the use of the translated ZmMsg.re,
ZmMsg.fw, and ZmMsg.fwd messages for when the client
"cleans up" the subject line. It does this by removing
all occurrences of "Re:" before pre-pending "Re:" (in
the case of a reply). This avoids subject lines like
"Re: Re: Re: Re: Re: Something".

Now the code will clean up occurrences of either "Re:"
or ZmMsg.re (in the case of reply) from the subject line.
This should cover the case where other mail clients use
localized versions of "Re:", etc, as well as the default.

http://bugzilla.zimbra.com/show_bug.cgi?id=30938