bug: 19702 - Trash/junk lifetime  based on the time message was moved to Trash or Junk
Added a new account/COS attribute called zimbraMailPurgeUseChangeDateForTrash.
When the value is set to FALSE, behavior is the same as before.  Messages are purged
from Trash based on the date they were added to the mailbox.  When the value is
TRUE, messages in Trash are purged based on the last date they were changed,
which will be the date they were moved to trash in most cases.

The mail_item.change_date column is currently not indexed.  I decided not to add
an index for the time being.  My tests concluded that most people
have on the order of hundreds or thousands of messages in the trash, so the
index on (mailbox_id, folder_id) should be good enough.

http://bugzilla.zimbra.com/show_bug.cgi?id=19702