bug: 31949
It appears that using searchDirectory to get the cos breaks the overview panel cos Item listing and the cos view. fixed the overview panel and the cos view. However, when a new cos is added, the cos list view is not updated automatically. If the auto-update is important. We can have a separate bug for it.

http://bugzilla.zimbra.com/show_bug.cgi?id=31949