bug: 32312
for identity/signature/datasource, the only need to call Provisioning.getInstance() is for getting the owner account by id.    We could've saved a ref to the Account object instead of keeping an account id.   The downside of that approach is that the ref could point to an Acount object that holds stale data if the account is expired from cache and then got modified and re-cached.  It's not likely to happen because identity/signature/datasource always have to be accessed via their Account object, so it's unlikely that the ref to the Account obj is expired, but it could happen; and this will break the current behavior.

the second solution would be to pass a Provisioning object to the getAccount() method of  identity/signature/datasource.  Identity and Signature are OK with this approach, but DataSource has a truck load of methods that calls getAccount() and spans to many callsites of those methods.  All those methods are only meaningful on server side anyway, most of them are Mailbox related that should never be called from SoapProvisioning anyway.

so I took the third approach: disable(throw INVALID_REQUEST) the getAccount() method in SoapIdentity/SoapSignature/SoapDatasource and in the Exception text refer callers to the new method getAccount(Provisioning).  I've checked currently there is no call to  identity/signature/datasource getAccount() in SoapProvisioning or any Zimbra code that uses SoapProvisioning.   If any Soap clients out there do use it and we had taken solution 2, they will have to change anyway to use the new method getAccount(Provisioning).

http://bugzilla.zimbra.com/show_bug.cgi?id=32312