Integrate change 121663 from FRANKLIN to main
-------------
Bug: 31242 - grantee loses admin permissions on the UI
Fix problems stemming from assumption that
the first share in a mountpoint's list of shares
is the one belonging to the current user. In the
case where the user does not have admin rights,
that's fine because there will only be one share.
But if the user has admin rights, and a non-admin
share got accepted on the organizer by another
user before this one, the first share is not the
user's share. To fix this, figure out which share
belongs to the user and return it via
ZmOrganizer::getMainShare().

http://bugzilla.zimbra.com/show_bug.cgi?id=31242