bug: 32769
Better error-checking of destination address for user commands.  <IMSubscribeRequest> and <IMSendMessageRequest> do sanity-checks of the destination addresses and will throw an exception (im.INVALID_ADDRESS code) if the destination address is immediately not valid.

The IM UI should be updated to give a better experience when this happens (instead of just the bare exception dialog)

http://bugzilla.zimbra.com/show_bug.cgi?id=32769