bug: 29634
Fix problems with HTML-entity mapping (HTML entities are *not* allowed in XHTML-IM, so they need to be mapped to their numeric equivalent).  Also un-map the numeric entities to the friendly HTML versions when logging the chats as HTML in the Chats folder.

Added new HtmlEntityMapper class which handles mapping/unmapping of html entities.

http://bugzilla.zimbra.com/show_bug.cgi?id=29634