bug: 28348
Removed zimbraPrefCalendarAllowMethodlessInvite account/COS preference.  This setting was reverted to a localconfig value.  This allows us to skip ics parsing entirely during message analysis if the calendar MIME part doesn't have the expected form.  With the previous per-user preference, it's not possible to skip the step because message analysis is common for multiple users who may have different preferences.

http://bugzilla.zimbra.com/show_bug.cgi?id=28348