bug 10695 - allow list for zimbraSmtpHostname
Now properly handling multiple values for zimbraSmtpHostname.
When more than one value is specified, we randomly pick
an MTA from the list.

Updated JMSession.getSession() to create a new Session
object each time, instead of using a single shared one.  This
is less efficient, but makes the code simpler and removes the
need to restart the server when SMTP settings are changed.
We can revisit JavaMail session caching later if we see perf
degradations.

http://bugzilla.zimbra.com/show_bug.cgi?id=10695