Integrate change 139995 from FRANKLIN to main
-------------
Bug: 33482 - Drag and Drop functionality is not working for Contacts in ABook
Fix DnD of contact cards for PV.
There were a couple of problems. First, the ID for
the DnD proxy wasn't being generated correctly,
which had the effect that you'd only get the DnD
icon on the first try for a particular contact.
For contact cards in the PV skin, text selection was
left enabled. That had the side effect that during
capture, the target reported for the event when the
mouse moved outside the cards view was always the
cards view, so the DnD icon stayed red. Text selection
has been disabled via CSS, as it is for the base skin.

http://bugzilla.zimbra.com/show_bug.cgi?id=33482