bug: 28550
p4 integrate -b FRANKLIN_to_main @140385,140385
there are two possible scenarios with bad addresses.  if the server accepts the SendMsg requests and subsequently fails to deliver to one of the addresses, server will deliver to valid addresses and bounce to sender Inbox indicating it wasn't delivered to the bad address.  if the server synchronously rejects the SendMsg request with SEND_ABORTED_ADDRESS_FAILURE, Desktop will bounce the message to Drafts folder and put a local only notification in user's Inbox.  The notification will clearly state that none of the recipients will receive the message.

http://bugzilla.zimbra.com/show_bug.cgi?id=28550