Integrate change 141288 from FRANKLIN to main
-------------
Bug: 33482 - Drag and Drop functionality is not working for Contacts in ABook
Code around strange IE6 behavior that breaks DnD
of contact cards.
The first problem was in sizing the DnD icon. That's
now done explicitly for IE6.
The other problem was somewhat weirder.
When the badge image (the thing that has the number
for multiple items, or a + for copy) is added to the icon
via appendChild, it blows away the content of the icon.
That only happens in IE6, and only for the contact card.
It was nothing obvious, so now we add the image
earlier in the code as HTML, which works.

http://bugzilla.zimbra.com/show_bug.cgi?id=33482