Integrate change 142118 from FRANKLIN to main
-------------
bug: 31965 - offline : Searching by 'Company Contacts' doesn't loads proper UI in address book
--
Because the contacts app is being activated for the first time, the list gets reset to the default Contacts folder. We avoid this now by checking if a current search has been made. Also fixed a bug with the "New Address Book" button not showing up in the current app toolbar. This was happening b/c the contacts app was not going thru activateApp via ZmZimbraMail.

http://bugzilla.zimbra.com/show_bug.cgi?id=31965