bug: 31694 - stream: MessageCache should not load full messages into memory
Updated MessageCache to only store MimeMessage objects
and not the blobs.  As of change 148087 FileBlobStore always
returns MimeMessage data as a BlobInputStream.  The blob
data is no longer read into memory, so there's no need to cache
it.  Updated MailItem.getContent() to get data directly from
StoreManager instead of going through the MessageCache.

Updated zimbraMessageCacheSize to specify the number of
items in the message cache, as opposed to the number of bytes.
The default is now 2000.  We'll need to add a 6.0 migration step
that updates this value for existing installations.

http://bugzilla.zimbra.com/show_bug.cgi?id=31694