bug: 25028
When account.zimbraCOSId changed , the inherited values are not reflected in the ModifyAccountResponse or in subsequest GetAccount/GetInfo calls until the account is expired from LRU cache.   This was because in Ldap LdapProvisioning.refreshEntry, we call Provisioning.getCos on the current entry and get a cached cos id for the account.  The default attr map(mDefaults) for the account entry is again set to the old cos after the account attrs are reloaded from LDAP.

FlushCache after MofifyAccount won't work either, becasue FlushCache calls refreshEntry, which does exactly the same as above.

The fix added a callback for zimbraCOSId; when zimbraCOSId is changed, the callback removes the account entry from cache.
Also, in ModifyAccount, do Provisioning.get(AccountBy) again after calling modifyAccount(it could be a different object now if the old entry was remvoe from cache), so it will definitely get the updated data and return them correctly in the response.

http://bugzilla.zimbra.com/show_bug.cgi?id=25028